+++ com.unity.cinemachine/CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - CmCamera is now CinemachineCamera.
 - ImputAxis refactor for recentering and momentary axis support.
 - Confiner2D and Confiner3D support smooth stop at bounds edge.
+- CinemachineIndependentImpulseListener renamed to CinemachineExternalImpulseListener.
 - Bugfix: Very occasional axis drift in SimpleFollow when viewing angle is +-90 degrees.